[FIX] web_editor, website: fix overlay handles display
A new parameter displayHandles was introduced on SnippetOptions to
display or not the handles on a snippet. If a snippet had no option
needing the handles overlay, the overlay (and handles) from its closest
parent with such option was.

Two use cases were not tackled correctly and should display overlay,
even with no option with displayHandles:
- root containers (which have no editable parent)
- inline snippets (movable snippets)

Part of https://github.com/odoo/odoo/pull/71909
task-2542318

X-original-commit: d811e92d166b27cafdc4e9594282b7b5dd25ed02