[IMP] project: do not display the archived sub-task in list
Currently, when go to project > settings > enable 'sub-tasks'
create a task >> add a sub-task and archive the sub-task
then the archived sub-task is still displayed in the sub-tasks list

so in this commit, removed the active_test context from the field to
hide the archived sub-task. also include the timesheet of archived
task while calculating the subtask_effective_hours and on click
of sub task hour open the timesheet of archived tasks.

PS: revert of the commit: https://github.com/odoo/odoo/commit/faaed797784ba13897195e282591a5171d619afd
As we would deduct archived tasks from the count and the number of planning hours,
and keep them included in the timesheets.

closes odoo/odoo#71719

Taskid: 2534898
X-original-commit: f3bffd67eb11372b935c319acf3794126b4874b5
Related: odoo/enterprise#18744
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>
Signed-off-by: Mahendra Barad <mba-odoo@users.noreply.github.com>