[FIX] project: ensure only stage_id field is used in ir_model_fields
Before this commit, the ir_model_fields used in the query could be
another field than the one expected, that is the stage_id field.
Indeed, since we don't check if the name column in the ir_model_fields
is right the stage_id field then we could have the user_id or another
many2one field.

This commit adds a check to be sure we only use the stage_id.

Related PR: #68343

closes odoo/odoo#71478

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>