[FIX] web_editor: improve Origins/07 background shape
The shape is a particular one which connects at the top and bottom with
two sub-shapes. Those two parts used the same color which made it
impossible for an user to control them individually. This is now the
case.

This allows for a theme override in themes which used them at the top
of their homepage to have the first part with the same color as the
header, and the second part with the same color as the next snippet.

Part of https://github.com/odoo/odoo/pull/68336
task-2491890

closes odoo/odoo#71768

X-original-commit: 8672912c2ebb2f65f844e454df6b4ec7610eedef
Related: odoo/design-themes#14
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>