[FIX] website: allow editing empty non-arch fields
08be49a897ce01275cfe7718a3cda0b34ce3b9a2 had added a rule that made
empty fields non-editable so that a snippetless website would not be
editable. This makes empty t-fields non-editable as well which is - to
say the least - not welcome.
This commit fixes that problem by making the rule more restrictive: what
we don't want to be able to edit are empty _arch_ field specifically.

closes odoo/odoo#71718

X-original-commit: cf52a091e620d78a7f3a6d4b0e211332e84e9eed
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>